Modified step 2
After cloning the repository I needed to cd into ./gsn before getting a list of all branches available. I added the 'cd gsn' instruction in step 2 of "How to create a local repository." This may not be in line with everyone else's experience but thought I would propose the change after running into it.